refactor class impl out of lib
